Refactoring the base code.
