+++ HenriksHobbyLager/Facades/IProductFacade.cs
@@ -1,10 +1,13 @@
-﻿namespace RefactoringExercise.Facades;
+﻿namespace HenriksHobbylager.Facades;
 
 internal interface IProductFacade
 {
+    // TODO: Change name om this to CreateProduct
     void AddProduct(string productName, int productQuantity, decimal productPrice);
     void DeleteProduct(int productId);
     void UpdateProduct(int productId, string productName, int productQuantity, decimal productPrice);
     Product SearchProduct(int productId);
     IEnumerable<Product> GetAllProducts();
+    
+    // TODO: Maby add IEnumerable<Product> SearchProducts(string searchTerm);
 }
