+++ HenriksHobbyLager/HenriksHobbyLager.csproj
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Include="Database\dummy.md" />
     <Compile Include="Interfaces\dummy.md" />
-    <Compile Include="Models\dummy.md" />
   </ItemGroup>
 
   <ItemGroup>
