+++ HenriksHobbyLager/Models/Product.cs
@@ -1,4 +1,4 @@
-﻿namespace RefactoringExercise.Models;
+﻿namespace HenriksHobbylager.Models;
 
 public class Product
 {
@@ -8,5 +8,5 @@ public class Product
     public int Stock { get; set; }
     public string Category { get; set; }
     public DateTime Created { get; set; }
-    public DateTime? LastUpdated { get; set; }  // Frågetecknet är för att jag är osäker på datumet
+    public DateTime? LastUpdated { get; set; }  
 }
