--- HenriksHobbyLager/Models/ShowAllProducts.cs
@@ -1,6 +1,0 @@
-﻿namespace RefactoringExercise.Models;
-
-public class ShowAllProducts
-{
-    
-}
