--- HenriksHobbyLager/Models/UpdateProduct.cs
@@ -1,6 +1,0 @@
-﻿namespace RefactoringExercise.Models;
-
-public class UpdateProduct
-{
-    
-}
