+++ HenriksHobbyLager/Repositories/Crud/DeleteProduct.cs
@@ -1,0 +1,19 @@
+﻿namespace HenriksHobbylager.Models;
+
+public class DeleteProduct
+{
+    Console.WriteLine("Skriv in produkt-id: ");
+    // TODO: Look this over.
+    // if (!int.TryParse(Console.ReadLine(), out int id))
+    // {
+    //     Console.WriteLine("Ogiltigt ID! Bara siffror är tillåtna här.");
+    //     return;
+    // }
+    var productId = int.Parse(Console.ReadLine());
+    // if (product == null)
+    // {
+    //     Console.WriteLine("Produkt hittades inte! Puh, inget blev raderat av misstag!");
+    //     return;
+    // }
+    _productFacade.DeleteProduct(productId);
+}
